requirements.txt
aniso8601==6.0.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
Flask==1.0.2
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
get==2019.3.22
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.6
pathlib==1.0.1
pkg-resources==0.0.0
post==2019.3.22
psycopg2==2.7.7
psycopg2-binary==2.7.7
public==2019.3.22
PyJWT==1.7.1
python-dotenv==0.10.1
pytz==2018.9
PyYAML==5.1
query-string==2019.3.22
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.15.1
